backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.0
celery[redis]==5.4.0
redis==5.1.0
boto3==1.35.0
openai==1.50.0
sendgrid==6.11.0
Pillow==10.4.0
python-dotenv==1.0.1
minio==7.2.0
pytest==8.3.0
pytest-asyncio==0.24.0
